requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
supabase
pyjwt
pillow
